Network.HTTP.Semantics.Server

HTTP server

type Server = Request -> Aux -> (Response -> [PushPromise] -> IO ()) -> IO ()

Server type. Server takes a HTTP request, should generate a HTTP response and push promises, then should give them to the sending function. The sending function would throw exceptions so that they can be logged.

The sending function must only be called once.

Request

data Request

Request from client.

Accessing request

requestMethod :: Request -> Maybe Method

Getting the method from a request.

requestPath :: Request -> Maybe Path

Getting the path from a request.

requestAuthority :: Request -> Maybe Authority

Getting the authority from a request.

requestScheme :: Request -> Maybe Scheme

Getting the scheme from a request.

requestHeaders :: Request -> TokenHeaderTable

Getting the headers from a request.

requestBodySize :: Request -> Maybe Int

Getting the body size from a request.

getRequestBodyChunk :: Request -> IO ByteString

Reading a chunk of the request body. An empty ByteString returned when finished.

getRequestBodyChunk' :: Request -> IO (ByteString, Bool)

Generalization of getRequestBodyChunk which also returns if the ByteString is the final one

getRequestTrailers :: Request -> IO (Maybe TokenHeaderTable)

Reading request trailers. This function must be called after getRequestBodyChunk returns an empty.

Aux

data Aux

Additional information.

auxTimeHandle :: Aux -> Handle

Time handle for the worker processing this request and response.

auxMySockAddr :: Aux -> SockAddr

Local socket address copied from Config.

auxPeerSockAddr :: Aux -> SockAddr

Remove socket address copied from Config.

Response

data Response

Response from server.

Creating response

responseNoBody :: Status -> ResponseHeaders -> Response

Creating response without body.

responseFile :: Status -> ResponseHeaders -> FileSpec -> Response

Creating response with file.

responseStreaming :: Status -> ResponseHeaders -> ((Builder -> IO ()) -> IO () -> IO ()) -> Response

Creating response with streaming.

responseBuilder :: Status -> ResponseHeaders -> Builder -> Response

Creating response with builder.

Generalized streaming interface

data OutBodyIface

Constructors

OutBodyIface

Fields

outBodyUnmask :: forall x. IO x -> IO x

Unmask exceptions in the thread spawned for the request body

This is used in the client: we spawn the new thread for the request body with exceptions masked, and provide the body of OutBodyStreamingIface with a callback to unmask them again (typically after installing an exception handler).

Unmasking in the server is a no-op, as here the scope of the thread that is spawned for the server is the entire handler, not just the response streaming body.

outBodyPush :: Builder -> IO ()

Push a new chunk

In http2, there is no direct correspondence between chunks and the resulting DATA frames sent: the chunks are collected (written to an internal write buffer) until we can fill a frame.

See also outBodyFlush.

outBodyPushFinal :: Builder -> IO ()

Push the final chunk

Using this function instead of outBodyPush can be used to guarantee that the final HTTP2 DATA frame is marked end-of-stream; with outBodyPush it may happen that an additional empty DATA frame is used for this purpose. Additionally, after calling this function, outBodyCancel will be a no-op.

outBodyCancel :: Maybe SomeException -> IO ()

Cancel the stream

Sends a RST_STREAM to the peer. If cancelling as the result of an exception, a Just should be provided which specifies the exception which will be stored locally as the reason for cancelling the stream; in this case, the error code sent with the RST_STREAM will be INTERNAL_ERROR (see https://datatracker.ietf.org/doc/html/rfc7540#section-7). If Nothing is given, the error code will be CANCEL.

If there is a partially constructed DATA frame at the time of cancellation, this frame is discarded. If this is undesirable, you should call outBodyFlush prior to cancelling.

outBodyFlush :: IO ()

Flush

This can be used to emit a DATA frame with the data collected so far (using outBodyPush), even if that DATA frame has not yet reached the maximum frame size. Calling outBodyFlush unnecessarily can therefore result in excessive overhead from frame headers.

If no data is available to send, this is a no-op.

responseStreamingIface :: Status -> ResponseHeaders -> (OutBodyIface -> IO ()) -> Response

Generalization of responseStreaming.

Accessing response

responseBodySize :: Response -> Maybe Int

Getter for response body size. This value is available for file body.

Trailers maker

type TrailersMaker = Maybe ByteString -> IO NextTrailersMaker

Trailers maker. A chunks of the response body is passed with Just. The maker should update internal state with the ByteString and return the next trailers maker. When response body reaches its end, Nothing is passed and the maker should generate trailers. An example:

{-# LANGUAGE BangPatterns #-}
import Data.ByteString (ByteString)
import qualified Data.ByteString.Char8 as C8
import Crypto.Hash (Context, SHA1) -- cryptonite
import qualified Crypto.Hash as CH

-- Strictness is important for Context.
trailersMaker :: Context SHA1 -> Maybe ByteString -> IO NextTrailersMaker
trailersMaker ctx Nothing = return $ Trailers [("X-SHA1", sha1)]
  where
    !sha1 = C8.pack $ show $ CH.hashFinalize ctx
trailersMaker ctx (Just bs) = return $ NextTrailersMaker $ trailersMaker ctx'
  where
    !ctx' = CH.hashUpdate ctx bs

Usage example:

let h2rsp = responseFile ...
    maker = trailersMaker (CH.hashInit :: Context SHA1)
    h2rsp' = setResponseTrailersMaker h2rsp maker

data NextTrailersMaker

Either the next trailers maker or final trailers.

Constructors

NextTrailersMaker TrailersMaker

Trailers [Header]

defaultTrailersMaker :: TrailersMaker

TrailersMake to create no trailers.

setResponseTrailersMaker :: Response -> TrailersMaker -> Response

Setting TrailersMaker to Response.

Push promise

data PushPromise

HTTP/2 push promise or sever push. Pseudo REQUEST headers in push promise is automatically generated. Then, a server push is sent according to promiseResponse.

Constructors

PushPromise

Fields

promiseRequestPath :: ByteString

Accessor for a URL path in a push promise (a virtual request from a server). E.g. "/style/default.css".

promiseResponse :: Response

Accessor for response actually pushed from a server.

pushPromise :: ByteString -> Response -> Int -> PushPromise

Creating push promise. The third argument is traditional, not used.

Types

type Path = ByteString

Path.

type Authority = String

Authority.

type Scheme = ByteString

"http" or "https".

data FileSpec

File specification.

Constructors

FileSpec FilePath FileOffset ByteCount

type FileOffset = Int64

Offset for file.

type ByteCount = Int64

How many bytes to read

type ReadN = Int -> IO ByteString

Reading n bytes.

defaultReadN :: Socket -> IORef (Maybe ByteString) -> ReadN

Naive implementation for readN.

NOTE: This function is intended to be used by a single thread only. (It is probably quite rare anyway to want concurrent reads from the same network socket.)

type PositionRead = FileOffset -> ByteCount -> Buffer -> IO ByteCount

Position read for files.

type PositionReadMaker = FilePath -> IO (PositionRead, Sentinel)

Making a position read and its closer.

data Sentinel

Constructors

Closer (IO ())
Closing a file resource. Its refresher is automatiaclly generated by the internal timer.

Refresher (IO ())
Refreshing a file resource while reading. Closing the file must be done by its own timer or something.

defaultPositionReadMaker :: PositionReadMaker

Position read based on Handle.